+++ webpack.common.js
@@ -21,6 +21,18 @@ module.exports = {
             {
                 test: /\.(png|svg|jpg|gif)$/u,
                 use:  'file-loader'
+            },
+            {
+                test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/u,
+                use:  [
+                    {
+                        loader:  'file-loader',
+                        options: {
+                            name:       '[name].[ext]',
+                            outputPath: 'fonts/'
+                        }
+                    }
+                ]
             }
         ]
     },